requirements.txt
cffi==1.15.1
cryptography==40.0.2
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pycparser==2.21
python-dotenv==1.0.0
